requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.1.0
bcrypt==3.2.0
certifi==2024.2.2
cffi==1.16.0
chargily_pay==2.1.3
charset-normalizer==3.3.2
click==8.1.7
elastic-transport==8.13.0
elasticsearch==8.13.0
flake8==7.0.0
flake8-bugbear==24.2.6
flake8-builtins==2.2.0
flake8-comprehensions==3.14.0
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
flask-restx==1.0.3
Flask-SQLAlchemy==3.0.2
fpdf==1.7.2
greenlet==3.0.1
gunicorn==22.0.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.0.1
Jinja2==3.1.3
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
Mako==1.2.3
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
packaging==24.0
pluggy==1.5.0
psutil==5.9.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.4.0
pytest==8.1.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
setuptools==68.2.2
six==1.16.0
SQLAlchemy==2.0.25
types-python-dateutil==2.9.0.20240316
types-requests==2.32.0.20240523
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==2.3.8
wheel==0.41.2
